refactor: 배팅 관련 객체 패키지 수정
