+++ src/test/java/participant/BettingMoneyTest.java
@@ -1,5 +1,6 @@
 package participant;
 
+import betting.BettingMoney;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
